Move TestAllocationLimits to new file
Do not run this test when the race detector is on.
